+++ Editor/UI/Framework/IssueTable.cs
@@ -230,6 +230,41 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
                 CellGUI(args.GetCellRect(i), args.item, args.GetColumn(i), ref args);
         }
 
+        public string GetCustomGroupPropertyCellString(IssueTableItem item, PropertyDefinition property)
+        {
+            string label = null;
+            var customPropertyIndex = PropertyTypeUtil.ToCustomIndex(property.type);
+            if (property.format == PropertyFormat.Bytes || property.format == PropertyFormat.Time || property.format == PropertyFormat.Percentage)
+            {
+
+                if (property.format == PropertyFormat.Bytes)
+                {
+                    ulong sum = 0;
+                    foreach (var childItem in item.children)
+                    {
+                        var issueTableItem = childItem as IssueTableItem;
+                        var value = issueTableItem.ProjectIssue.GetCustomPropertyUInt64(customPropertyIndex);
+                        sum += value;
+                    }
+
+                    label = Formatting.FormatSize(sum);
+                }
+                else
+                {
+                    float sum = 0;
+                    foreach (var childItem in item.children)
+                    {
+                        var issueTableItem = childItem as IssueTableItem;
+                        var value = issueTableItem.ProjectIssue.GetCustomPropertyFloat(customPropertyIndex);
+                        sum += value;
+                    }
+                    label = property.format == PropertyFormat.Time ? Formatting.FormatTime(sum) : Formatting.FormatPercentage(sum, 1);
+                }
+            }
+
+            return label;
+        }
+
         void CellGUI(Rect cellRect, TreeViewItem treeViewItem, int columnIndex, ref RowGUIArgs args)
         {
             var property = m_Layout.properties[columnIndex];
@@ -271,34 +306,10 @@ namespace Unity.ProjectAuditor.Editor.UI.Framework
                 }
                 else if (PropertyTypeUtil.IsCustom(property.type))
                 {
-                    var customPropertyIndex = PropertyTypeUtil.ToCustomIndex(propertyType);
-                    if (property.format == PropertyFormat.Bytes || property.format == PropertyFormat.Time || property.format == PropertyFormat.Percentage)
+                    string label = GetCustomGroupPropertyCellString(item, property);
+
+                    if(!string.IsNullOrEmpty(label))
                     {
-                        string label;
-                        if (property.format == PropertyFormat.Bytes)
-                        {
-                            ulong sum = 0;
-                            foreach (var childItem in item.children)
-                            {
-                                var issueTableItem = childItem as IssueTableItem;
-                                var value = issueTableItem.ProjectIssue.GetCustomPropertyUInt64(customPropertyIndex);
-                                sum += value;
-                            }
-
-                            label = Formatting.FormatSize(sum);
-                        }
-                        else
-                        {
-                            float sum = 0;
-                            foreach (var childItem in item.children)
-                            {
-                                var issueTableItem = childItem as IssueTableItem;
-                                var value = issueTableItem.ProjectIssue.GetCustomPropertyFloat(customPropertyIndex);
-                                sum += value;
-                            }
-                            label = property.format == PropertyFormat.Time ? Formatting.FormatTime(sum) : Formatting.FormatPercentage(sum, 1);
-                        }
-
                         EditorGUI.LabelField(cellRect, label, labelStyle);
                     }
                 }